+++ src/lib.rs
@@ -1,6 +1,5 @@
 #![no_std]
 
-pub mod error;
 pub mod reader;
 pub mod writer;
 pub mod types;