+++ CHANGELOG.md
@@ -6,3 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.0] - 2026-09-07
+
+### Added
+
+- initial driver impl
+- half-duplex example
+- full-duplex example
+
+### Changed
+
+- renamed `Reading` to `GpsReading`
+- renamed `AdafruitUltimateGPS` to `AdafruitUltimateGps`
+- moved `GpsReading` from `reader` to `types`
+- moved `GpsError` from `error` to `types`